ADD: Dashboard ad its pages routes
