+++ src/routes/+layout.svelte
@@ -2,6 +2,7 @@
 	import type { LayoutData } from './$types';
 	import '../app.css';
 	import { Truck } from 'lucide-svelte';
+	import { page } from '$app/stores';
 
 	let { data, children }: { data: LayoutData; children: any } = $props();
 </script>
@@ -15,17 +16,6 @@
 					<h1><Truck size={18} style="vertical-align: -2px; margin-right: 6px;" /> Fleet Manager</h1>
 				</div>
 				
-				<nav class="main-nav">
-					<a href="/dashboard" class="nav-link">Dashboard</a>
-					<a href="/records" class="nav-link">Records</a>
-					<a href="/vehicles" class="nav-link">Vehicles</a>
-					<a href="/analytics" class="nav-link">Analytics</a>
-					
-					{#if data.user.role === 'super_admin' || data.user.role === 'org_admin'}
-						<a href="/admin" class="nav-link">Admin</a>
-					{/if}
-				</nav>
-
 				<div class="user-menu">
 					<div class="user-info">
 						<span class="user-name">{data.user.first_name} {data.user.last_name}</span>
@@ -43,8 +33,24 @@
 		</main>
 	</div>
 {:else}
-	<!-- Public layout (login page) -->
-	{@render children()}
+	<!-- Public layout -->
+	<div class="public-container">
+		{#if !$page.url.pathname.startsWith('/auth')}
+		<header class="public-header">
+			<div class="public-inner">
+				<a class="brand" href="/">Fleet Manager</a>
+				<nav class="public-nav">
+					<a href="/blog" class="public-link">Blog</a>
+					<a href="/contact" class="public-link">Contact</a>
+					<a href="/auth/login" class="public-cta">Sign in</a>
+				</nav>
+			</div>
+		</header>
+		{/if}
+		<main class="public-main">
+			{@render children()}
+		</main>
+	</div>
 {/if}
 
 <style>
@@ -81,26 +87,6 @@
 		margin: 0;
 		font-size: 1.25rem;
 		font-weight: 700;
-		color: #1f2937;
-	}
-
-	.main-nav {
-		display: flex;
-		gap: 0.5rem;
-		flex: 1;
-	}
-
-	.nav-link {
-		padding: 0.5rem 1rem;
-		color: #4b5563;
-		text-decoration: none;
-		font-weight: 500;
-		border-radius: 6px;
-		transition: all 0.2s;
-	}
-
-	.nav-link:hover {
-		background: #f3f4f6;
 		color: #1f2937;
 	}
 
@@ -153,6 +139,27 @@
 		padding: 2rem 1rem;
 	}
 
+	/* Public layout styles */
+	.public-header {
+		background: white;
+		border-bottom: 1px solid #e5e7eb;
+	}
+	.public-inner {
+		max-width: 1200px;
+		margin: 0 auto;
+		padding: 0.75rem 1rem;
+		display: flex;
+		align-items: center;
+		justify-content: space-between;
+	}
+	.brand { color: #111827; font-weight: 700; text-decoration: none; }
+	.public-nav { display: flex; gap: 0.75rem; align-items: center; }
+	.public-link { color: #4b5563; text-decoration: none; padding: 0.4rem 0.6rem; border-radius: 6px; }
+	.public-link:hover { background: #f3f4f6; color: #1f2937; }
+	.public-cta { color: white; background: #2563eb; padding: 0.45rem 0.8rem; border-radius: 6px; text-decoration: none; font-weight: 600; }
+	.public-cta:hover { background: #1d4ed8; }
+	.public-main { max-width: 1200px; margin: 0 auto; padding: 2rem 1rem; }
+
 	@media (max-width: 768px) {
 		.header-content {
 			flex-wrap: wrap;
@@ -160,29 +167,6 @@
 			padding: 0.75rem 1rem;
 		}
 
-		.logo h1 {
-			font-size: 1rem;
-		}
-
-		.main-nav {
-			order: 3;
-			width: 100%;
-			flex-wrap: wrap;
-			gap: 0.25rem;
-		}
-
-		.nav-link {
-			font-size: 0.875rem;
-			padding: 0.375rem 0.75rem;
-		}
-
-		.user-info {
-			display: none;
-		}
-
-		.logout-btn {
-			font-size: 0.8rem;
-			padding: 0.375rem 0.75rem;
-		}
+		.logo h1 { font-size: 1rem; }
 	}
 </style>
