+++ src/routes/blog/+page.svelte
@@ -1,0 +1,19 @@
+<script lang="ts">
+  // Public Blog landing placeholder
+</script>
+
+<svelte:head>
+  <title>Blog - Fleet Manager</title>
+  <meta name="description" content="News and updates about Fleet Manager." />
+</svelte:head>
+
+<section class="hero">
+  <h1>Blog</h1>
+  <p>Stories, updates and engineering notes. Coming soon.</p>
+</section>
+
+<style>
+  .hero { max-width: 900px; margin: 2rem auto; padding: 0 1rem; }
+  h1 { margin: 0 0 0.5rem 0; color: #111827; }
+  p { color: #4b5563; }
+</style>
