+++ src/routes/dashboard/vehicles/+page.server.ts
@@ -1,3 +1,4 @@
+// src/routes/dashboard/vehicles/+page.server.ts
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ locals }) => {
